[refactor]: rename Home to Shows
