+++ src/views/routes.js
@@ -1,0 +1,13 @@
+import { Routes, Route } from 'react-router-dom';
+
+import Home from './Shows';
+
+const App = () => {
+  return (
+    <Routes>
+      <Route path="/" element={<Home />}></Route>
+    </Routes>
+  );
+};
+
+export default App;
